Remove unnecessary inline query result interfaces

+++ src/Telegram.Bot/Types/InlineQueryResults/InputMessageContent/InputLocationMessageContent.cs
@@ -1,10 +1,12 @@
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 
+// ReSharper disable once CheckNamespace
 namespace Telegram.Bot.Types.InlineQueryResults
 {
     /// <summary>
-    /// Represents the content of a location message to be sent as the result of an <see cref="InlineQuery">inline query</see>.
+    /// Represents the content of a location message to be sent as the result of an
+    /// <see cref="InlineQuery">inline query</see>.
     /// </summary>
     [JsonObject(MemberSerialization.OptIn, NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
     public class InputLocationMessageContent : InputMessageContent
@@ -13,13 +15,13 @@ namespace Telegram.Bot.Types.InlineQueryResults
         /// Latitude of the location in degrees
         /// </summary>
         [JsonProperty(Required = Required.Always)]
-        public float Latitude { get; }
+        public double Latitude { get; }
 
         /// <summary>
         /// Longitude of the location in degrees
         /// </summary>
         [JsonProperty(Required = Required.Always)]
-        public float Longitude { get; }
+        public double Longitude { get; }
 
         /// <summary>
         /// Optional. The radius of uncertainty for the location, measured in meters; 0-1500
@@ -40,7 +42,8 @@ namespace Telegram.Bot.Types.InlineQueryResults
         public int? Heading { get; set; }
 
         /// <summary>
-        /// Optional. Maximum distance for proximity alerts about approaching another chat member, in meters. For sent live locations only.
+        /// Optional. Maximum distance for proximity alerts about approaching another chat member,
+        /// in meters. For sent live locations only.
         /// </summary>
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
         public int? ProximityAlertRadius { get; set; }
@@ -50,7 +53,7 @@ namespace Telegram.Bot.Types.InlineQueryResults
         /// </summary>
         /// <param name="latitude">The latitude of the location</param>
         /// <param name="longitude">The longitude of the location</param>
-        public InputLocationMessageContent(float latitude, float longitude)
+        public InputLocationMessageContent(double latitude, double longitude)
         {
             Latitude = latitude;
             Longitude = longitude;
